Develop (#204)
* Fixes SickRage/SickRage#1873

* Added BBC Scotland timezone (#166)

Currently when you add a TV Show which is from "BBC Scotland" like for example "Limmy's Show" we don't know what timezone to put it in. This adds "BBC Scotland" to the correct time zone, which is "Europe/London".

* Add IRIB TV Iran and duplicate Yomiuri TV Japan (#201)

Yomiuri TV has JP suffix here but there's no show using that according to theTVDB. Duplicating without JP suffix which then matches theTVDB which has one show.

Adding IRIB which was only added to the local version previously.

* Adding exception for Flash (2014) (#200)

Flash (2014) is published as The Flash 2014 on torrent sites.